Move the resolving logic out of client
